requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
lsf
python-dotenv
aiohttp
